+++ src/main/java/com/avp/common/entity/living/human/marine/MarineSpawning.java
@@ -1,8 +1,32 @@
 package com.avp.common.entity.living.human.marine;
 
+import com.avp.AVP;
+import com.avp.common.block.AVPBlockTags;
+import com.avp.common.config.AVPConfig;
+import com.avp.common.entity.living.human.AbstractHumanMob;
+import net.minecraft.world.entity.SpawnPlacements;
+
 public class MarineSpawning {
 
-    // public static final SpawnPlacements.SpawnPredicate<MarineMob> PREDICATE = AlienSpawning.createPredicate(
-    // AVP.config.spawnConfigs.MARINE_SPAWN
-    // );
+     public static final SpawnPlacements.SpawnPredicate<MarineMob> PREDICATE = createPredicate(
+            AVP.config.spawnConfigs.MARINE_SPAWN
+     );
+
+    public static <T extends AbstractHumanMob> SpawnPlacements.SpawnPredicate<T> createPredicate(
+            AVPConfig.SpawnConfigs.SpawnSettings container
+    ) {
+        return (
+                entityType,
+                serverLevelAccessor,
+                mobSpawnType,
+                blockPos,
+                randomSource
+        ) -> {
+            var belowState = serverLevelAccessor.getBlockState(blockPos.below());
+            var spawnableBlock = AVPBlockTags.MARINE_SPAWN_BLOCKS;
+            var isValidPos = belowState.is(spawnableBlock);
+
+            return isValidPos;
+        };
+    }
 }